requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.12
pydantic-settings==2.8.1
email-validator==2.2.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
